+++ WebApp/Areas/Admin/Controllers/AccountController.cs
@@ -7,6 +7,8 @@ using System.Net;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
+using System.Web.Security;
+using WebApp.Areas.Admin.ViewModels;
 using WebApp.Models;
 using WebApp.Utils;
 using WebApp.ViewModels;
@@ -14,40 +16,110 @@ using WebApp.ViewModels;
 namespace WebApp.Areas.Admin.Controllers
 {
     [Authorize(Roles = Role.Admin)]
-    public class AccountController : BaseAccountController
+    public class AccountController : Controller
     {
-        public AccountController() : base() { }
-
-        public AccountController(ApplicationUserManager userManager, ApplicationSignInManager signInManager)
-            : base(userManager, signInManager) { }
-
-        protected override void SetManagedRoles()
+        private readonly ApplicationDbContext _context;
+        private ApplicationSignInManager _signInManager;
+        private ApplicationUserManager _userManager;
+        public ApplicationSignInManager SignInManager
         {
-            roles.Add(Role.Staff);
-            roles.Add(Role.Trainer);
+            get
+            {
+                return _signInManager ?? HttpContext.GetOwinContext().Get<ApplicationSignInManager>();
+            }
+            private set
+            {
+                _signInManager = value;
+            }
+        }
+
+        public ApplicationUserManager UserManager
+        {
+            get
+            {
+                return _userManager ?? HttpContext.GetOwinContext().GetUserManager<ApplicationUserManager>();
+            }
+            private set
+            {
+                _userManager = value;
+            }
+        }
+        public AccountController()
+        {
+            _context = new ApplicationDbContext();
+
         }
 
         [HttpGet]
         public async Task<ActionResult> Index()
         {
-            var model = new List<GroupedUsersViewModel>();
+            var staffRole = await _context.Roles.SingleOrDefaultAsync(r => r.Name == Role.Staff);
+            var trainerRole = await _context.Roles.SingleOrDefaultAsync(r => r.Name == Role.Trainer);
 
-            foreach (var roleName in roles)
+            var model = new GroupedUsersViewModel()
             {
-                var role = await RoleManager.FindByNameAsync(roleName);
-
-                var groupedUsers = new GroupedUsersViewModel()
-                {
-                    Type = roleName,
-                    Users = await _context.Users.Where(x => x.Roles.Select(y => y.RoleId).Contains(role.Id)).ToListAsync(),
-                };
-                model.Add(groupedUsers);
-            }
+                Staffs = await _context.Users.Where(x => x.Roles.Select(y => y.RoleId).Contains(staffRole.Id)).ToListAsync(),
+                Trainers = await _context.Users.Where(x => x.Roles.Select(y => y.RoleId).Contains(trainerRole.Id)).ToListAsync(),
+                Trainees = null
+            };
 
             return View(model);
         }
+        [HttpGet]
+        public ActionResult Create()
+        {
+            AccountRegisterViewModel model = new AccountRegisterViewModel()
+            {
+                Roles = new List<string> { Role.Staff, Role.Trainer }
+            };
+            return View(model);
+        }
 
-        protected override async Task<UserViewModel> LoadUserViewModel(string userId)
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<ActionResult> Create(AccountRegisterViewModel model)
+        {
+            if (ModelState.IsValid)
+            {
+                var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
+                var result = await UserManager.CreateAsync(user, model.Password);
+
+
+                if (result.Succeeded)
+                {
+                    await UserManager.AddToRoleAsync(user.Id, model.Role);
+                    switch (model.Role)
+                    {
+                        case Role.Trainer:
+                            var profile = new Trainer()
+                            {
+                                UserId = user.Id,
+                                Specialty = null,
+                            };
+                            _context.Trainers.Add(profile);
+                            await _context.SaveChangesAsync();
+                            break;
+                        default:
+                            break;
+                    }
+                    return RedirectToAction("Index", "Account");
+                }
+                AddErrors(result);
+            }
+
+            // If we got this far, something failed, redisplay form
+            return View(model);
+        }
+
+        private void AddErrors(IdentityResult result)
+        {
+            foreach (var error in result.Errors)
+            {
+                ModelState.AddModelError("", error);
+            }
+        }
+
+        private async Task<UserViewModel> LoadUserViewModel(string userId)
         {
             var user = await UserManager.FindByIdAsync(userId);
 
@@ -62,56 +134,177 @@ namespace WebApp.Areas.Admin.Controllers
                 Roles = new List<string>(roles)
             };
 
-            model = await LoadUserProfile(model);
-
-            return model;
-        }
-
-        protected override async Task<UserViewModel> LoadUserProfile(UserViewModel model)
-        {
-            if (roles.Contains(Role.Trainer))
+            if (roles.Any(r => r == Role.Trainer))
             {
-                var trainer = await _context.Trainers.SingleOrDefaultAsync(u => u.UserId == model.User.Id);
-
-                if (trainer == null)
-                {
-                    trainer = new Models.Trainer()
-                    {
-                        UserId = model.User.Id,
-                        Specialty = null
-                    };
-
-                    _context.Trainers.Add(trainer);
-                    await _context.SaveChangesAsync();
-                }
+                var trainer = await _context.Trainers.SingleOrDefaultAsync(u => u.UserId == userId);
                 model.Specialty = trainer.Specialty;
             }
 
             return model;
         }
 
-        protected override async Task<UserViewModel> UpdateUserProfile(UserViewModel model)
+        public async Task<ActionResult> Details(string id)
         {
-            if (roles.Contains(Role.Trainer))
+            if (id == null)
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+            UserViewModel model = await LoadUserViewModel(id);
+
+            if (model == null)
+                return HttpNotFound();
+            return View(model);
+        }
+
+        [HttpGet]
+        public async Task<ActionResult> Delete(string id, bool? saveChangesError = false)
+        {
+            var user = await UserManager.FindByIdAsync(id);
+
+            if (user == null)
             {
-                var trainer = await _context.Trainers.SingleOrDefaultAsync(u => u.UserId == model.User.Id);
-
-                if (trainer == null)
-                {
-                    trainer = new Models.Trainer()
-                    {
-                        UserId = model.User.Id,
-                        Specialty = model.Specialty
-                    };
-                    _context.Trainers.Add(trainer);
-                }
-                else
-                    trainer.Specialty = model.Specialty;
-
-                await _context.SaveChangesAsync();
+                return HttpNotFound();
             }
 
-            return model;
+            var model = new UserViewModel()
+            {
+                User = user,
+                Roles = new List<string>(await UserManager.GetRolesAsync(user.Id))
+            };
+
+            if (saveChangesError == true)
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
+            }
+
+            return View(model);
         }
+
+        [HttpPost, ActionName("Delete")]
+        public async Task<ActionResult> ConfirmedDelete(string id)
+        {
+            var user = await UserManager.FindByIdAsync(id);
+
+            if (user == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            IdentityResult result = await UserManager.DeleteAsync(user);
+
+            if (result.Succeeded)
+            {
+                return RedirectToAction(nameof(Index), "Account");
+            }
+
+            return RedirectToAction(nameof(Delete), new { id, saveChangesError = true });
+        }
+
+        [HttpGet]
+        public async Task<ActionResult> Edit(string id)
+        {
+            if (id == null)
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+            UserViewModel model = await LoadUserViewModel(id);
+
+            if (model == null)
+                return HttpNotFound();
+            return View(model);
+        }
+
+        [HttpPost]
+        public async Task<ActionResult> Edit(UserViewModel model)
+        {
+            if (ModelState.IsValid)
+            {
+                var user = model.User;
+                var userinDb = await UserManager.FindByIdAsync(user.Id);
+
+                if (userinDb == null)
+                    return HttpNotFound();
+                userinDb.FullName = user.FullName;
+                userinDb.Age = user.Age;
+                userinDb.Address = user.Address;
+                userinDb.Email = user.Email;
+                userinDb.UserName = user.Email;
+
+                IdentityResult result = await UserManager.UpdateAsync(userinDb);
+
+                if (model.Specialty != null)
+                {
+                    var profile = await _context.Trainers.SingleOrDefaultAsync(p => p.UserId == userinDb.Id);
+                    profile.Specialty = model.Specialty;
+                }
+                await _context.SaveChangesAsync();
+
+                if (result.Succeeded)
+                    return RedirectToAction(nameof(Index));
+                else
+                    AddErrors(result);
+            }
+
+            // If we got this far, something failed, redisplay form
+            return View(model);
+        }
+
+        [HttpGet]
+        public ActionResult ResetPassword(string email = null)
+        {
+            if (email == null)
+                return View();
+
+            var model = new ResetPasswordViewModel()
+            {
+                Email = email
+            };
+
+            return View(model);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<ActionResult> ResetPassword(ResetPasswordViewModel model)
+        {
+            if (!ModelState.IsValid)
+            {
+                return View(model);
+            }
+            var user = await UserManager.FindByEmailAsync(model.Email);
+
+            if (user == null)
+            {
+                ViewBag.ErrorMessage = "The user does not exist";
+                return View(model);
+            }
+
+            var roles = await UserManager.GetRolesAsync(user.Id);
+            if (!roles.All(r => r == Role.Staff || r == Role.Trainer))
+            {
+                ViewBag.ErrorMessage = "The user cannot be reset. Permission is denied.";
+                return View(model);
+            }
+
+            model.Code = await UserManager.GeneratePasswordResetTokenAsync(user.Id);
+            IdentityResult result = await UserManager.ResetPasswordAsync(user.Id, model.Code, model.Password);
+
+            if (result.Succeeded)
+            {
+                return RedirectToAction(nameof(ResetPasswordConfirmation), "Account", new { email = user.Email });
+            }
+
+            AddErrors(result);
+            return View();
+        }
+
+        [Route("{email}")]
+        [HttpGet]
+        public ActionResult ResetPasswordConfirmation(string email)
+        {
+            ViewBag.Email = email;
+            return View();
+        }
+
     }
 }
