+++ WebApp/Areas/Admin/Views/Account/Delete.cshtml
@@ -1,4 +1,5 @@
 ﻿@model WebApp.ViewModels.UserViewModel
+@using WebApp.Utils;
 @{ 
 	ViewBag.Title = "Delete Account";
 }
