+++ WebApp/Areas/Admin/Views/Account/Details.cshtml
@@ -1,10 +1,10 @@
-@model WebApp.ViewModels.UserViewModel
-@{ 
-	ViewBag.Title = "Edit profile";
+﻿@model WebApp.ViewModels.UserViewModel
+@{
+	ViewBag.Title = "Account Details";
 }
 
-@Html.Partial("UpdateProfilePartial")
-
-<div>
+@Html.Partial("DetailsProfilePartial", Model)
+<p>
+	@Html.ActionLink("Edit Profile", "Edit", new { id = Model.User.Id }) |
 	@Html.ActionLink("Back to List", "Index")
-</div>
+</p>
