+++ WebApp/Areas/Admin/Views/Account/Index.cshtml
@@ -1,48 +1,72 @@
-@model ICollection<WebApp.ViewModels.GroupedUsersViewModel>
+﻿@model WebApp.ViewModels.GroupedUsersViewModel
 @{
 	ViewBag.Title = "Dashboard";
 	int i = 0;
 }
 <h2>Dashboard</h2>
 
-<ul class="list-group">
-	<li class="list-group-item">
-		@Html.ActionLink("Create new account", "Create", "Account")
-	</li>
-	<li class="list-group-item">
-		@Html.ActionLink("Reset Password", "ResetPassword")
+<div class="list-group">
+	@Html.ActionLink("Create new account", "Create", null, htmlAttributes: new { @class = "list-group-item list-group-item-action" })
+	@Html.ActionLink("Reset Password", "ResetPassword", null, htmlAttributes: new { @class = "list-group-item list-group-item-action" })
+</div>
 
-	</li>
-</ul>
-
-@foreach (var groupedUsers in Model)
-{
-	<h3>@groupedUsers.Type Accounts</h3>
-	<table class="table table-striped">
-		<thead>
+<h3>Staff Accounts</h3>
+<table class="table table-striped">
+	<thead>
+		<tr>
+			<th scope="col">#</th>
+			<th scope="col">UserName</th>
+			<th scope="col">Full Name</th>
+			<th scope="col">Action</th>
+		</tr>
+	</thead>
+	<tbody>
+		@foreach (var item in @Model.Staffs)
+		{
+			{ i++; }
 			<tr>
-				<th scope="col">#</th>
-				<th scope="col">UserName</th>
-				<th scope="col">Full Name</th>
-				<th scope="col">Action</th>
+				<th scope="row">@i</th>
+				<td>@item.UserName</td>
+				<td>@item.FullName</td>
+				<td>
+					@Html.ActionLink("Delete", "Delete", new { id = item.Id }, new { type = "button", @class = "btn btn-danger" })
+					@Html.ActionLink("Edit profiles", "Edit", new { id = item.Id }, new { @class = "btn btn-success" })
+					@Html.ActionLink("Details", "Details", new { id = item.Id }, new { @class = "btn btn-info" })
+					@Html.ActionLink("Reset Password", "ResetPassword", new { email = item.Email }, new { @class = "btn btn-danger" })
+				</td>
 			</tr>
-		</thead>
-		<tbody>
-			@foreach (var user in groupedUsers.Users)
-			{
-				{ i++; }
-				<tr>
-					<th scope="row">@i</th>
-					<td>@user.UserName</td>
-					<td>@user.FullName</td>
-					<td>
-						@Html.ActionLink("Delete", "Delete", new { id = user.Id }, new { @class = "btn btn-danger" })
-						@Html.ActionLink("Edit profiles", "Edit", new { id = user.Id }, new { @class = "btn btn-success" })
-						@Html.ActionLink("Details", "Details", new { id = user.Id }, new { @class = "btn btn-info" })
-						@Html.ActionLink("Reset Password", "ResetPassword", new { email = user.Email }, new { @class = "btn btn-danger" })
-					</td>
-				</tr>
-			}
-		</tbody>
-	</table>
-}
+		}
+
+	</tbody>
+</table>
+
+<h3>Trainer Accounts</h3>
+<table class="table table-striped">
+	<thead>
+		<tr>
+			<th scope="col">#</th>
+			<th scope="col">UserName</th>
+			<th scope="col">Full Name</th>
+			<th scope="col">Action</th>
+		</tr>
+	</thead>
+	<tbody>
+		@{ i = 0; }
+		@foreach (var item in @Model.Trainers)
+		{
+			{ i++; }
+			<tr>
+				<th scope="row">@i</th>
+				<td>@item.UserName</td>
+				<td>@item.FullName</td>
+				<td>
+					@Html.ActionLink("Delete", "Delete", new { id = item.Id }, new { @class = "btn btn-danger" })
+					@Html.ActionLink("Edit profiles", "Edit", new { id = item.Id }, new { @class = "btn btn-success" })
+					@Html.ActionLink("Details", "Details", new { id = item.Id }, new { @class = "btn btn-info" })
+					@Html.ActionLink("Reset Password", "ResetPassword", new { email = item.Email }, new { @class = "btn btn-danger" })
+				</td>
+			</tr>
+		}
+
+	</tbody>
+</table>
