+++ WebApp/Areas/Staff/Views/Trainee/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebApp.ViewModels.AccountRegisterViewModel
+﻿@model WebApp.Areas.Admin.ViewModels.AccountRegisterViewModel
 @{
 	ViewBag.Title = "Create Trainee Account";
 }
