+++ WebApp/Areas/Staff/Views/Trainee/Delete.cshtml
@@ -4,7 +4,7 @@
 	ViewBag.Title = "Delete Account";
 }
 
-@Html.Partial("DetailsProfilePartial")
+@Html.Partial("DetailsProfilePartial", Model)
 
 @using (Html.BeginForm("Delete", "Trainee", FormMethod.Post))
 {
