+++ WebApp/Areas/Staff/Views/Trainee/Edit.cshtml
@@ -3,7 +3,7 @@
 	ViewBag.Title = "Edit profile";
 }
 
-@Html.Partial("UpdateProfilePartial")
+@Html.Partial("UpdateProfilePartial", Model)
 
 <div>
 	@Html.ActionLink("Back to List", "Index")
