--- WebApp/Areas/Trainer/Views/Trainer/Details.cshtml
@@ -1,9 +1,0 @@
-﻿@model WebApp.ViewModels.UserViewModel
-@{
-    ViewBag.Title = "Account Details";
-}
-@Html.Partial("DetailsProfilePartial")
-<p>
-    @Html.ActionLink("Edit", "EditAccount", new { id = Model.User.Id }) |
-    @Html.ActionLink("Back to List", "Index")
-</p>
