--- WebApp/Areas/Trainer/Views/Trainer/Edit.cshtml
@@ -1,10 +1,0 @@
-﻿@model WebApp.ViewModels.UserViewModel
-@{
-    ViewBag.Title = "Edit profile";
-}
-
-@Html.Partial("UpdateProfilePartial")
-
-<div>
-    @Html.ActionLink("Back to List", "Index")
-</div>
