--- WebApp/Areas/Trainer/Views/Trainer/Index.cshtml
@@ -1,8 +1,0 @@
-﻿@model PagedList.IPagedList<WebApp.Models.ApplicationUser>
-@using PagedList.Mvc
-@{
-    ViewBag.Title = "Trainee Management";
-    int i = 0;
-}
-<h2>@Html.ActionLink((string)ViewBag.Title, "Index")</h2>
-
