--- WebApp/Areas/Trainer/Views/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "~/Areas/Trainer/Views/Shared/_Layout.cshtml";
-}
