+++ WebApp/ViewModels/GroupedUsersViewModel.cs
@@ -8,7 +8,11 @@ namespace WebApp.ViewModels
 {
     public class GroupedUsersViewModel
     {
-        public string Type { get; set; }
-        public List<ApplicationUser> Users { get; set; }
+        public List<ApplicationUser> Staffs { get; set; }
+
+        public List<ApplicationUser> Trainers { get; set; }
+
+        public List<ApplicationUser> Trainees { get; set; }
+
     }
 }
