--- WebApp/Views/Manage/UpdateProfile.cshtml
@@ -1,8 +1,0 @@
-﻿@model WebApp.ViewModels.UserViewModel
-@using WebApp.Utils;
-
-@Html.Partial("UpdateProfilePartial")
-
-<div>
-    @Html.ActionLink("Back to List", "Index")
-</div>
