--- WebApp/Views/Manage/ViewProfile.cshtml
@@ -1,11 +1,0 @@
-﻿@model WebApp.ViewModels.UserViewModel
-@using WebApp.Utils;
-
-@Html.Partial("DetailsProfilePartial")
-
-<div>
-    @Html.ActionLink("Back to List", "Index")
-</div>
-
-
-
