+++ WebApp/Views/Shared/UpdateProfilePartial.cshtml
@@ -1,98 +1,85 @@
-@model WebApp.ViewModels.UserViewModel
+﻿@model WebApp.ViewModels.UserViewModel
 @using WebApp.Utils;
 
-@using (Html.BeginForm())
+@using (Html.BeginForm(null, null, FormMethod.Post, htmlAttributes: new { @class = "" }))
 {
 	@Html.AntiForgeryToken()
 
-	<div class="form-horizontal">
+	<div class="container">
 		<h4>Update Profile</h4>
 		<hr />
 		@Html.ValidationSummary(true, "", new { @class = "text-danger" })
 		@Html.HiddenFor(model => model.User.Id)
 
-		<div class="form-group">
-			@Html.LabelFor(model => model.User.UserName, htmlAttributes: new { @class = "control-label col-md-2" })
-			<div class="col-md-10">
+		<div class="mb-3">
+			@Html.LabelFor(model => model.User.UserName, htmlAttributes: new { @class = "form-label" })
+			<div>
 				@Html.EditorFor(model => model.User.UserName, new { htmlAttributes = new { @class = "form-control", @readonly = "readonly" } })
 				@Html.ValidationMessageFor(model => model.User.UserName, "", new { @class = "text-danger" })
 			</div>
 		</div>
 
-		<div class="form-group">
-			@Html.LabelFor(model => model.User.FullName, htmlAttributes: new { @class = "control-label col-md-2" })
-			<div class="col-md-10">
+		<div class="mb-3">
+			@Html.LabelFor(model => model.User.FullName, htmlAttributes: new { @class = "form-label" })
+			<div>
 				@Html.EditorFor(model => model.User.FullName, new { htmlAttributes = new { @class = "form-control" } })
 				@Html.ValidationMessageFor(model => model.User.FullName, "", new { @class = "text-danger" })
 			</div>
 		</div>
 
-		<div class="form-group">
-			@Html.LabelFor(model => model.User.Age, htmlAttributes: new { @class = "control-label col-md-2" })
-			<div class="col-md-10">
+		<div class="mb-3">
+			@Html.LabelFor(model => model.User.Age, htmlAttributes: new { @class = "form-label" })
+			<div class="col-form-label">
 				@Html.EditorFor(model => model.User.Age, new { htmlAttributes = new { @class = "form-control" } })
 				@Html.ValidationMessageFor(model => model.User.Age, "", new { @class = "text-danger" })
 			</div>
 		</div>
 
-		<div class="form-group">
-			@Html.LabelFor(model => model.User.Address, htmlAttributes: new { @class = "control-label col-md-2" })
-			<div class="col-md-10">
-				@Html.EditorFor(model => model.User.Address, new { htmlAttributes = new { @class = "form-control" } })
-				@Html.ValidationMessageFor(model => model.User.Address, "", new { @class = "text-danger" })
-			</div>
+		<div class="mb-3">
+			@Html.LabelFor(model => model.User.Address, htmlAttributes: new { @class = "form-label" })
+
+			@Html.EditorFor(model => model.User.Address, new { htmlAttributes = new { @class = "form-control" } })
+			@Html.ValidationMessageFor(model => model.User.Address, "", new { @class = "text-danger" })
 		</div>
 
-		<div class="form-group">
-			@Html.LabelFor(model => model.User.Email, htmlAttributes: new { @class = "control-label col-md-2" })
-			<div class="col-md-10">
-				@Html.EditorFor(model => model.User.Email, new { htmlAttributes = new { @class = "form-control" } })
-				@Html.ValidationMessageFor(model => model.User.Email, "", new { @class = "text-danger" })
-			</div>
+		<div class="mb-3">
+			@Html.LabelFor(model => model.User.Email, htmlAttributes: new { @class = "form-label" })
+			@Html.EditorFor(model => model.User.Email, new { htmlAttributes = new { @class = "form-control" } })
+			@Html.ValidationMessageFor(model => model.User.Email, "", new { @class = "text-danger" })
 		</div>
 
-		<div class="form-group">
-			@Html.LabelFor(model => model.Roles, htmlAttributes: new { @class = "control-label col-md-2" })
-			<div class="col-md-10">
-				@Html.EditorFor(model => model.Roles, new { htmlAttributes = new { @class = "form-control", @value = string.Join(", ", Model.Roles), @readonly = "readonly" } })
-				@Html.ValidationMessageFor(model => model.User.Email, "", new { @class = "text-danger" })
-			</div>
+		<div class="mb-3">
+			@Html.LabelFor(model => model.Roles, htmlAttributes: new { @class = "form-label" })
+			@Html.EditorFor(model => model.Roles, new { htmlAttributes = new { @class = "form-control", @value = string.Join(", ", Model.Roles), @readonly = "readonly" } })
+			@Html.ValidationMessageFor(model => model.User.Email, "", new { @class = "text-danger" })
 		</div>
 
 		@if (Model.Roles.Any(r => r == Role.Trainer))
 		{
-			<div class="form-group">
-				@Html.LabelFor(model => model.Specialty, htmlAttributes: new { @class = "control-label col-md-2" })
-				<div class="col-md-10">
-					@Html.EditorFor(model => model.Specialty, new { htmlAttributes = new { @class = "form-control" } })
-					@Html.ValidationMessageFor(model => model.Specialty, "", new { @class = "text-danger" })
-				</div>
+			<div class="mb-3">
+				@Html.LabelFor(model => model.Specialty, htmlAttributes: new { @class = "form-label" })
+				@Html.EditorFor(model => model.Specialty, new { htmlAttributes = new { @class = "form-control" } })
+				@Html.ValidationMessageFor(model => model.Specialty, "", new { @class = "text-danger" })
 			</div>
 		}
 
 		@if (Model.Roles.Any(r => r == Role.Trainee))
 		{
-			<div class="form-group">
-				@Html.LabelFor(model => model.BirthDate, htmlAttributes: new { @class = "control-label col-md-2" })
-				<div class="col-md-10">
-					@Html.EditorFor(model => model.BirthDate, new { htmlAttributes = new { @class = "form-control" } })
-					@Html.ValidationMessageFor(model => model.BirthDate, "", new { @class = "text-danger" })
-				</div>
+			<div class="mb-3">
+				@Html.LabelFor(model => model.BirthDate, htmlAttributes: new { @class = "form-label" })
+				@Html.EditorFor(model => model.BirthDate, new { htmlAttributes = new { @class = "form-control" } })
+				@Html.ValidationMessageFor(model => model.BirthDate, "", new { @class = "text-danger" })
 			</div>
 
-			<div class="form-group">
-				@Html.LabelFor(model => model.Education, htmlAttributes: new { @class = "control-label col-md-2" })
-				<div class="col-md-10">
-					@Html.EditorFor(model => model.Education, new { htmlAttributes = new { @class = "form-control" } })
-					@Html.ValidationMessageFor(model => model.Education, "", new { @class = "text-danger" })
-				</div>
+			<div class="mb-3">
+				@Html.LabelFor(model => model.Education, htmlAttributes: new { @class = "form-label" })
+				@Html.EditorFor(model => model.Education, new { htmlAttributes = new { @class = "form-control" } })
+				@Html.ValidationMessageFor(model => model.Education, "", new { @class = "text-danger" })
 			</div>
 		}
 
-		<div class="form-group">
-			<div class="col-md-offset-2 col-md-10">
-				<input type="submit" value="Save changes" class="btn btn-default" />
-			</div>
+		<div class="input-group">
+			<input type="submit" value="Save changes" class="btn btn-primary" />
 		</div>
 	</div>
 }
